refactor: auth directory and update .env variable name
