+++ src/features/auth/auth.module.ts
@@ -2,13 +2,12 @@ import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { KakaoApiClient } from './kakao-api-client';
 import { User } from 'src/entities/user.entity';
-import { KakaoStrategy } from './kakao-strategy';
+import { KakaoStrategy } from './strategy/kakao-strategy';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { UserModule } from '../user/user.module';
-import { JwtStrategy } from './jwt-strategy';
+import { JwtStrategy } from './strategy/jwt-strategy';
 
 @Module({ 
   imports: [
@@ -22,7 +21,7 @@ import { JwtStrategy } from './jwt-strategy';
 
   ],
   controllers: [AuthController],
-  providers: [AuthService, KakaoApiClient, KakaoStrategy, JwtStrategy],
+  providers: [AuthService, KakaoStrategy, JwtStrategy],
   exports: [PassportModule, AuthService]
 })
 export class AuthModule {}
